+++ client/components/dashboard/dashboard.spec.js
@@ -1,18 +1,19 @@
-/* eslint-disable no-unused-vars */
 import { shallow } from 'enzyme';
-import TestUtils from 'react-addons-test-utils';
 import React from 'react';
-import Dashboard from './index';
 
-import {expect} from 'chai';
+import Dashboard from './index';
+import MonthlySummary from './monthlySummary';
+
+import { expect } from 'chai';
 
 describe('<Dashboard />', () => {
 
-  it('Main dashboard page includes a heading of "Welcome to the Merchant Dashboard"', () => {
+  it('should contain a <MonthlySummary /> component on the dashboard', () => {
 
-    const expected = 'Welcome to the Merchant Dashboard';
+    const expected = 1;
     const wrapper = shallow(<Dashboard />);
-    const actual = wrapper.find('h1').props().children;
+    const actual = wrapper.find(MonthlySummary)
+      .length;
 
     expect(actual).equals(expected);