+++ client/components/dashboard/index.js
@@ -1,14 +1,13 @@
-/* eslint-disable max-len */
 import React from 'react';
+
+import MonthlySummary from './monthlySummary';
 
 class Dashboard extends React.Component {
 
   render() {
     return (
-      <div className="text-center m-t-lg">
-        <h1>
-          Welcome to the Merchant Dashboard
-        </h1>
+      <div className="m-t-lg">
+        <MonthlySummary />
         <small>
           This is an application skeleton while it's being built out.  Many changes to come, keep getting latest!
         </small>